+++ core/src/main/resources/reference.conf
@@ -598,7 +598,7 @@ services {
   }
 
   CostCatalogService {
-    class = "cromwell.services.cost.CostCatalogService"
+    class = "cromwell.services.cost.GcpCostCatalogService"
     config {
       catalogExpirySeconds = 86400
     }
