+++ services/src/test/scala/cromwell/services/cost/GcpCostCatalogServiceSpec.scala
@@ -12,23 +12,23 @@ import org.scalatest.matchers.should.Matchers
 import java.time.Duration
 import java.io.{FileInputStream, ObjectInputStream}
 
-object CostCatalogServiceSpec {
-  val catalogExpirySeconds: Long = 1 // Short so we can do a cache expiry test
-  val config = ConfigFactory.parseString(s"catalogExpirySeconds = $catalogExpirySeconds")
-  val mockTestDataFilePath = "services/src/test/scala/cromwell/services/cost/serializedSkuList.testData"
+object GcpCostCatalogServiceSpec {
+  val catalogExpirySeconds: Long = 1 // Short duration so we can do a cache expiry test
+  val config: Config = ConfigFactory.parseString(s"catalogExpirySeconds = $catalogExpirySeconds")
+  val mockTestDataFilePath: String = "services/src/test/scala/cromwell/services/cost/serializedSkuList.testData"
   def loadMockDataFromFile: Iterable[Sku] = {
-    val fis = new FileInputStream(CostCatalogServiceSpec.mockTestDataFilePath)
+    val fis = new FileInputStream(GcpCostCatalogServiceSpec.mockTestDataFilePath)
     val ois = new ObjectInputStream(fis)
     ois.readObject().asInstanceOf[List[Sku]]
   }
 }
 
-class CostCatalogServiceTestActor(serviceConfig: Config, globalConfig: Config, serviceRegistry: ActorRef)
-    extends CostCatalogService(serviceConfig, globalConfig, serviceRegistry) {
-  override def fetchNewCatalog: Iterable[Sku] = CostCatalogServiceSpec.loadMockDataFromFile
+class GcpCostCatalogServiceTestActor(serviceConfig: Config, globalConfig: Config, serviceRegistry: ActorRef)
+    extends GcpCostCatalogService(serviceConfig, globalConfig, serviceRegistry) {
+  override def fetchNewCatalog: Iterable[Sku] = GcpCostCatalogServiceSpec.loadMockDataFromFile
 }
 
-class CostCatalogServiceSpec
+class GcpCostCatalogServiceSpec
     extends TestKitSuite
     with AnyFlatSpecLike
     with Matchers
@@ -36,15 +36,15 @@ class CostCatalogServiceSpec
     with ImplicitSender {
   behavior of "CostCatalogService"
 
-  def constructTestActor: CostCatalogService =
+  def constructTestActor: GcpCostCatalogService =
     TestActorRef(
-      new CostCatalogService(CostCatalogServiceSpec.config, CostCatalogServiceSpec.config, TestProbe().ref)
+      new GcpCostCatalogService(GcpCostCatalogServiceSpec.config, GcpCostCatalogServiceSpec.config, TestProbe().ref)
     ).underlyingActor
   private val testActorRef = constructTestActor
   private val minimumExpectedSkus = 10
 
   it should "properly load mock test data file" in {
-    CostCatalogServiceSpec.loadMockDataFromFile.toList.length shouldBe >(minimumExpectedSkus)
+    GcpCostCatalogServiceSpec.loadMockDataFromFile.toList.length shouldBe >(minimumExpectedSkus)
   }
 
   it should "cache catalogs properly" in {
@@ -56,7 +56,7 @@ class CostCatalogServiceSpec
     )
 
     val freshActor = constructTestActor
-    val shortDuration: Duration = Duration.ofSeconds(CostCatalogServiceSpec.catalogExpirySeconds)
+    val shortDuration: Duration = Duration.ofSeconds(GcpCostCatalogServiceSpec.catalogExpirySeconds)
 
     // Sanity check that the catalog has been initially fetched with a new timestamp
     freshActor.getSku(testLookupKey) // do a lookup to trigger any lazy loading
